Extracted serde code for Container into separate file
