+++ src/prodbg/viewdock/src/lib.rs
@@ -286,43 +286,6 @@ mod test {
 //    }
 //
 //    #[test]
-//    fn test_container_serialize_0() {
-//        let container_in = Container {
-//            docks: Vec::new(),
-//            rect: Rect::new(4.0, 5.0, 2.0, 8.0)
-//        };
-//
-//        let serialized = serde_json::to_string(&container_in).unwrap();
-//        let container_out: Container = serde_json::from_str(&serialized).unwrap();
-//
-//        assert_eq!(container_out.docks.len(), 0);
-//        // expect that rect is not serialized and set to zero
-//        assert_eq!(container_out.rect.x as i32, 0);
-//        assert_eq!(container_out.rect.y as i32, 0);
-//        assert_eq!(container_out.rect.width as i32, 0);
-//        assert_eq!(container_out.rect.height as i32, 0);
-//    }
-//
-//    #[test]
-//    fn test_container_serialize_1() {
-//        let container_in = Container {
-//            docks: vec![Dock {
-//                handle: DockHandle(1),
-//                plugin_name: "registers".to_owned(),
-//                plugin_data: Some(vec!["some_data".to_owned(), "more_data".to_owned()]),
-//                rect: Rect::new(4.0, 5.0, 2.0, 8.0)
-//            }],
-//            rect: Rect::default(),
-//        };
-//
-//        let serialized = serde_json::to_string(&container_in).unwrap();
-//        let container_out: Container = serde_json::from_str(&serialized).unwrap();
-//
-//        assert_eq!(container_out.docks.len(), 1);
-//        assert_eq!(container_out.docks[0].plugin_name, "registers");
-//    }
-//
-//    #[test]
 //    fn test_split_serialize_0() {
 //        let split_in = Split {
 //            left: None,
